nodegit: if not exists the plugin folder then clone it
