+++ .gitignore
@@ -10,5 +10,5 @@ storage/
 backend/mobile/upload/
 config/
 dist/
-
+playground
 webpack
