+++ package.json
@@ -16,7 +16,7 @@
     "start": "nodemon ./backend/mobile",
     "local": "nodemon ./backend/mobile",
     "start:dev": "nodemon ./backend/mobile development",
-    "safe": "node ./index.js --mode=safemode",
+    "safe": "node ./backend/mobile/index.js --mode=safemode",
     "digital-signage": "node ./index.js --config=./config/digital-signage.json",
     "start:test": "export PATH_ENV=https://raw.githubusercontent.com/anhoev/cms_configs/master/digital_signage.json && node ./backend/mobile/index.js",
     "submodule-update": "git submodule foreach -q --recursive 'git checkout $(git config -f $toplevel/.gitmodules submodule.$name.branch || echo master) && git pull'",
@@ -63,6 +63,7 @@
     "request": "^2.88.0",
     "sass.js": "^0.10.13",
     "signale": "^1.4.0",
+    "simple-git": "^1.110.0",
     "socket.io": "^2.2.0",
     "socket.io-client": "^2.2.0",
     "traverse": "^0.6.6",
